Overwrite empty namespace py3
Summary:
# Problem
{D80489335} revealed that the codemod currently adds a duplicate namespace py3 in the presence of a `namespace py3 ""`, which results in an invalid thrift file.

# The solution
Replace an empty `namespace py3` with the supplied namespace.

# The rationale
The pre-requisite for a thrift file to be part of the units of work is that the replacement namespace is non-empty.

Unless explicitly chosen by the user, Thrift files should not have an empty `namespace py3`. If the user explicitly selected an empty `namespace py3`, then that `thrift_library`, and the `thrift_files` that it contains, will not be part of the codemod units of work.

This implies that the codemod must encounter an empty `namespace py3` IFF the thrift file contains an empty `namespace py3` NOT selected by the user.

For e.g. in the case of {D80489335} that revealed this problem, {D49909728}, which was part of a ThriftPackage codemod, added the `namespace py3 ""` . This means there is no risk if we replace that `namespace py3 ""` with the `namespace py3 configerator.structs.admarket.dpa`.

Reviewed By: aristidisp

Differential Revision: D80569824

fbshipit-source-id: 0d840837227de9875b5de2fc00e52438d760ff8f