Enable resource pools via all opt-out/opt-in flags if process is being migrated.
Summary:
When checking if resource pool migration is enabled, we will modify all opt-out and opt-in flags to reflect that resource pools should be used.

This is immediately useful for migrating `unicorn` services, which all opt out of resource pools via the `--thrift_disable_resource_pools` flag.

The `isMigrationEnabled()` also checks the return value of the existing `useResourcePoolsFlagsSet()` function after modifying the opt-out/opt-in flags and ensures migration doesn't proceed if that function has already returned false (resource pools are disabled).

Reviewed By: tlj77

Differential Revision: D80303479

fbshipit-source-id: a9fe9f205a1e2fd0a77a602c0ade351ebac3161f