Remove short circuit logic
Summary: This removes the short circuit logic by calling each function separately. Each function has a side effect of potentially updating a set that contains the dependent types of the program. By skipping a function with short circuit logic we will have an incomplete set when a `map` contains dependent types for both the key and value types.

Reviewed By: dtolnay

Differential Revision: D80509996

fbshipit-source-id: 65452528761ea25a9209b34322a03b64dadda716