Do not rely on TypeFinder
Summary:
Previously we have `TypeFinder` that only adds nested schema from the struct we are pretty printing.

This doesn't work well with Any since Any might require a schema that's unrelated to the original struct.

Since diff switched to `getSyntaxGraphDefinitionNodeByUri` so that we can get schema from uri directly.

After this diff, TypeFinder is not useful anymore and we can delete it later (though let's do in a separate diff since it requires some refactoring).

This diff made TypeFinder stateless, next diff we will completely delete it.

Reviewed By: thedavekwon

Differential Revision: D81016234

fbshipit-source-id: e3a5d01a78ba6b9ff2dfe91b3ba1a7cb3aee874f