createBiDiClientCallback
Summary:
Since the eventual client-facing interface for BiDi will be client sink + client stream we want to reuse as much of the existing plumbing as possible, including their associated bridges.
This diff introduces a class for interoperating stream/sink client callbacks with bidi server callback, proxying events between the two contracts. `DeletionGuard` is introduced because the return values from the stream/sink callback methods provide an incomplete view into the overall state of the bidi request and so the latter must be computed out-of-band.

Reviewed By: sazonovkirill

Differential Revision: D81055843

fbshipit-source-id: 95d337193ec28c50574fb54dc90dfd9bbd223f20