Lint for resolution conflicts
Summary:
Add a lint for ambiguous usage of unscoped identifiers. The global resolution logic for thrift squashes all identifiers into a single namespace, where the name of the thrift file is prepended as a scope to all its definitions.

When resolving identifiers, this allows an identifier in a program to implicitly add a scope name equivalent to the current program's name, e.g.
```
> in other/foo.thrift
struct Foo {}

> in foo.thrift
include "b/foo.thrift"

struct Bar {
    1: Foo foo;  // <-- This is resolved as foo.Foo, which points to b/foo.thrift
}
```

Reviewed By: praihan

Differential Revision: D81287755

fbshipit-source-id: 98d802a4edeb6687ca86a96c3062ec87b0f3a840