Add logging for when union is constructed with multiple fields
Summary:
We currently have logging in `ThriftUnionSerializationTrait.php` if a union is serialized/deserialized with multiple fields. There is, however, no logging if a union is directly *constructed* with multiple fields. In such cases, we are interested if the extra fields (every field that is not the last defined field in the IDL), are accessed. If they are not accessed, then it is safe to migrate the union to use the strict-unions implementation.

This diff adds the `logIncorrectFieldAccessed()` method which will be used in updated union codegen in D79062399.

In addition, changed the existing `signal_log_in_psp` usage to be more structured instead of having long string keys.

Reviewed By: rmakheja

Differential Revision: D78720958

fbshipit-source-id: 734e6a5c5a31712156db77dfa6acaf285bfb856a