Remove backdoors from all over the codebase behind IIF (test change)
Summary:
I'm structuring this change in two parts:

1) Changes to control side of the experiment -- adding the IsItFaster, but with the test branch identical to control to see what changes in the control side. (D81806419)

2) Changes to the test side of the experiment -- modifying the test branch to show what changes in test. (This diff)

The idea here is to ultimately remove these backdoors wherever possible, and this will likely happen regardless of the results. I also want to test this with IIF, as backdoors also have a latency/CPU cost.

This accounts for probably ~1/4th of all backdoor usages across the codebase.

Reviewed By: puffpio

Differential Revision: D81806418

fbshipit-source-id: 7429a9d61f4c23132bbd2e3a00a7cd2e6e6bcf06