create interface to switch between refactored rocketserverconnections
Summary:
To support refactoring the RocketServerConnection there needs to be an interface so we can swap in and out different implementations of the RocketServerConnection.

Currently usings 'I' prefix for the inferface - that will be removed later after the classes are renamed.

This diff just lands the interface and it will be used in later diffs.

Reviewed By: tlj77

Differential Revision: D82180007

fbshipit-source-id: f8ba0ee4caf5c72549936f73d3b11a3414721ff9