Remove Adapter from conformance test
Summary: adapter is a language dependent feature that we shouldn't be testing in conformance test especially in round-trip-test. We can argue that it tests whether user can write a adapter that conform to same wire protocol, but it pollutes the conformance test too much. This removes lines nonconforming.txt from 12260 -> 3219 (~74%)

Reviewed By: praihan

Differential Revision: D82756065

fbshipit-source-id: c843f2950da82040742ed43aa0cbc40504f3b1c8