Fix build_templates
Summary:
It seems that a recent macos toolchain change (T238287798) is enforcing -Wno-string-concatentation more aggressively. Specifically, the compiler is suspicious about string literals being concatentated in an array.

This change seems to be breaking ///thrift/compiler/generate:templates for macos builds (T237713850)

The fix I've chosen is to parenthesize individual array elements, so that concatentation across elements is not possible. Seems to make the compiler happy.

Apparently the concatenation exists because Windows has a max limit on literal size, so there's no changing that.

Reviewed By: iahs

Differential Revision: D82771128

fbshipit-source-id: fd7aec60fb948e02218de1a6c264a713b969cbf8