Add fbthrift::*::serialize_ref support for users to serialize a Thrift object via a reference instead of taking by value
Summary:
This diff adds a new method for Thrift serialization in Rust that doesn't require the struct being serialized to be accepted by value - instead, a shared (immutable) reference should be sufficient.

It's technically possible to do this today with the existing pieces but there isn't a top-level user-friendly set of functions that can do this.

Reviewed By: emersonford

Differential Revision: D83061207

fbshipit-source-id: ccff08a0eff44a0dcc2f1866e4aa340df17f541e