Simplify public surface of ServerGeneratorStreamBuilder
Summary: We do not like abstracting `TwoWayBridge` methods behind the bridge's interface, as we find brige's interface to be more confusing than the `TwoWayBridge`'s one. This also open the door for the future refactoring where we want to use the bridge as a data structure and separate the code that uses the bridge from the bridge itself, replacing inheritance with composition. This feels like a good idea because we now have instances where the same bridge (data structure) can be used differently in different scenarios.

Reviewed By: evanjzou

Differential Revision: D82696783

fbshipit-source-id: af434fa94532b4f851eefac759351fce50725390