Add 'object.is?' to Whisker standard library
Summary:
Add function `object.is?` for templates to check the type of a Whisker object.

This will allow templates to differentiate between mstch (Whisker map) and Whisker prototype (native handle) contexts for incremental migration, where templates may need to handle rendering in the context of a mstch node or the equivalent Whisker prototype handle.

Reviewed By: iahs

Differential Revision: D83294699

fbshipit-source-id: 1217f6c13ebfc346911a8f43b4b684337a34aee6