Fix richcomp ops for Immutable Thrift Python Sets
Summary:
This fixes bug in implementation of rich comparison for immutable Thrift Python Sets which were used in issubset and issuperset.

Main bug is here that for Sets,`x > y` or `x >= y` can not be implemented using `not x<=y` or `not x < y`. Because < for Sets form partial orders unlike Integers which form total order.
For example `x > y` was implemented as `not x <= y` => `not ( x == y or x < y)`. Counter example for this, for `x={0}` and `y={1}`. `x > y` is `False`, and `not (x == y or x < y)` is `True`.

This is documented in Python documentation (https://docs.python.org/3.10/library/stdtypes.html?highlight=issubset#frozenset.copy):
{F1982273794}

Reviewed By: aristidisp

Differential Revision: D82980041

fbshipit-source-id: 218f0a0c6de015f5aee356f8b9e18b1468dfeb4a