Change `__repr__` to produce valid, evaluable code
Summary:
In immutable containers, we include `i`  in the __repr__ method to support immutability and hashability. We initially added`i`  to mutable containers to maintain consistency with the immutable ones.

This diff moves `i` out of mutable containers to prevent confusion and to address Python type issues reported by users.

https://fb.workplace.com/groups/thriftpydeprecatedunification/posts/2229258100922462

#internal
build-all

Reviewed By: prakashgayasen

Differential Revision: D83067179

fbshipit-source-id: fd22db304c24db30ef386cc8bfcc219393c9b283