Stop ResourcePool::stop() from spamming logs if someone calls it in a loop.
Summary:
This stops logging in `ResourcePool::stop()` from spamming logs if someone calls it in a loop.

I chose to allow the first 10 log messages to go through every minute for the following reasons.
1. Multiple resource pools may hit this problem. With 10 log messages, we're more likely to see messages from different resource pools that fail to stop at the same time.
2. In cases where the process is calling this in a loop, it'll be obvious that they're doing that, but we won't completely fill their logs with spam. They'll get maximum 10 logs per minute.

### **Note**

The motivation for this change is that the `ricardo` client hit this in the past and had to disable resource pools to avoid the log spam (see D58223168). I'm unable to reproduce the failure in `ResourcePool::stop()` after reverting that change. Regardless, we'd rather not repeat the log spam.

Reviewed By: tlj77

Differential Revision: D83397339

fbshipit-source-id: 31bf81cd62f7d3753a27fb461c380406e3b71fe7