Throw more informative exception on unset union field access
Summary:
Addresses a user pain point that it is hard to debug unset union field access exceptions because they don't contain any information about the union or field being accessed.
Uses existing vtable mechanism to provide the information with marginal build/runtime cost. It is possible to add another member to the vtable to provide field names instead of ids, but it would likely require additional codegen in order to avoid circular dependency problems and most of the value is already captured by this implementation.

Reviewed By: vitaut

Differential Revision: D84002800

fbshipit-source-id: 98c378d942cd7fa75e65296a96d826ba5075dba5