Add string-based container benchmarks in preparation for Map optimization implementation
Summary:
This diff adds benchmarks for string-based lists and sets, in addition to the existing integer-based ones. The results show that immutable integer containers are **~40x** faster than immutable string containers for first random access, due to special optimizations for integer types.

This suggests there are opportunities to further improve the performance of string-based containers.

Reviewed By: yoney

Differential Revision: D84196046

fbshipit-source-id: 8214975cbf615eaf7fd24e61e73f48f050df3559