Add null check for interactions
Summary:
Add a null check to return an exception instead of crash if the interaction is in a bad state.

Meta:
This is a mitigation for S557239. We only expect this to occur in ServiceLab environments where interaction replay is enabled. Unfortunately we don't yet understand the root cause, so this is a temporary fix to try and unblock the failing workflows for now.

Reviewed By: sazonovkirill

Differential Revision: D84178261

fbshipit-source-id: b7dc8b04edcd4a45ddb8247acec7d4d7f04cb858