Remove impossible condition
Summary:
I am 99.999% confident this condition is impossible.

Returning `nil` `conn` and `nil` `error` from the `Accept()` call would be EXTREMELY non-idiomatic.

This code was first added in D60283615 - I think perhaps the condition was needed for the `default` case - which at the time was not a `return` statement. But with the state of the code today - the condition is impossible.

Reviewed By: siutsin

Differential Revision: D84547092

fbshipit-source-id: 0b6ec330d3b1f5a3e10f5083bfbb0aee6cd9b70b