Add protocol_methods fallback to op::encode
Summary: Customizing protocol_methods or Cpp2Ops was the legacy way to customize serialization for a type before adapters. Add a fallback to op::encode to call this when regular serialization doesn't work to unblock the rollout while remaining types are migrated. Also adds a fallback to direct conversion for cases where the custom serialization is even more implicit.

Reviewed By: praihan

Differential Revision: D83698413

fbshipit-source-id: c3cbe1d8513f3e5fdfe3e05d5aa02cd468bb45d8