Make SyntaxGraph::asTypeSystem replace AnyStruct with built-in Any
Summary:
`TypeSystem` has first-class support for `Any` so we should make this conversion happen as early as possible.

Note that we still populate the `TypeSystem` `StructNode` for `AnyStruct`. So manually looking up the struct will still work. We just change all the `TypeRef`s to the built-in `Any` so no edges in the graph will actually point to the struct.

Reviewed By: iahs

Differential Revision: D84929548

fbshipit-source-id: 061814a5297f15f2306ade50caf4c11c8ee31ae4