Fix contextprop propagation on the HTTP receive path
Summary:
This is a followup from D84114293, with one more case that weren't handling.

When we make requests to the thrift server using the HTTP transport, we aren't
fully propagating the thrift framework metadata. The rocket transport will send
this metadata fully inline, while the HTTP transport sends it via a header, and
the default hooks consume / delete the header. The rocket server does a similar
thing of preserving the `IOBuf` before creating the `ThriftRequest` object and
then later copying it into the `RequestContext` object.

I don't _love_ this implementation, since it adds some extra b64 work, but I
think we would all prefer to kill the HTTP transport instead :)

Reviewed By: michaelneu

Differential Revision: D85102770

fbshipit-source-id: 90daa21ff9e6de87dcfc5dff224897ce39c8058e