www changes for Handling stream service fatal exceptions
Summary:
Client reads the TApplicationException message from the exception metadata which is expected to contain the deserialized message. To have access to deserialized fatal exception message, we need to use a new error handler specific to streaming and send it directly via the extension instead of alite controller.

This  diff adds the www changes.
 - Add a new error page that is used only if stream has started
 - The new error page uses streaming extension to send the fatal exception
 - In addition, catch any exception while stream is active and send it using the extension instead of alite controller.

Reviewed By: rikeoka

Differential Revision: D83917936

fbshipit-source-id: c934dab4341645cf183cf0ba3bd2e3078ba3e035