Sink exception handling
Summary: This diff fixes some of the sink exception handling and adds a unit test for throwing a declared exception from the sink. We reuse the same pattern from the sink implementation - the `SinkElementDecoder` will handle the exception based on type, so we only need to `co_yield` the `co_result`.

Reviewed By: sazonovkirill

Differential Revision: D84551467

fbshipit-source-id: 5b1e336448bf68b88eb519fd78680fa3d032b963