Fix bug with racy request
Summary:
ReconnectingRequestChannel tries to queue requests when the channel is
not yet connected, and does it in this sequence:
1. Request comes in
2. New channel is created
3. Request is queued
4. Eventually callback that channel is connected, and queue is drained

For additional requests that come in between 3 and 4 the expectation is that
these requests are also queued. However, the invariant in code is
`isChannelGood()`, but this relies on the implementation to tell us if it's
connected or not.

For Rocket, this doesn't work as the connection is good as soon as it's created
(https://fburl.com/code/18gwzhnl). So when new request comes in, we think it's
already connected and skip the queuing logic.

The fix is to instead just rely on `isReconnecting_`, which is our internal
invariant, and is set after 4.

Reviewed By: tlj77

Differential Revision: D85554985

fbshipit-source-id: 81a9235c2122f358d35804355b91a6c6a54660af